nav bar routes completed
