+++ src/App.css
@@ -6,7 +6,9 @@ html{
 #root{
   background-image: url("https://wallpaperaccess.com/full/130220.jpg")
 }
-
+#text{
+  color: white;
+}
 /* Style the navbar */
 #navbar {
   overflow: hidden;
