+++ src/components/Nav.js
@@ -6,9 +6,9 @@ class Nav extends React.Component {
     render() {
       return (
         <nav id="navbar">
-            <a href="#home">Home</a>
-            <a href="#news">News</a>
-            <a href="#collection">Collection</a>
+            <a href="/">Home</a>
+            <a href="/about">About</a>
+            <a href="/collection">Collection</a>
         </nav> 
 
       );
