+++ src/pages/About.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+class About extends React.Component {
+
+    render() {
+      return (
+        <div className='about-page'>
+          <div id="text">
+              THIS IS MY ABOUT PAGE
+          </div>
+        </div>
+      );
+    }
+}
+
+export default About;
