+++ src/pages/Home.js
@@ -1,0 +1,14 @@
+import React from 'react'
+
+class Home extends React.Component {
+
+  render() {
+    return (
+      <div id="text">
+          THIS IS MY Home PAGE
+      </div>
+    );
+  }
+}
+
+export default Home;
